# backend/full_openapi.json
{"openapi":"3.1.0","info":{"title":"Moderation System Backend","version":"0.1.0"},"paths":{"/users":{"get":{"summary":"List Users","operationId":"list_users_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Response List Users Users Get"}}}}}},"post":{"summary":"Create User","operationId":"create_user_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts":{"post":{"summary":"Create Post","description":"Creates a new post and dispatches a moderation task.","operationId":"create_post_posts_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Posts","operationId":"list_posts_posts_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Post"},"title":"Response List Posts Posts Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/{post_id}/moderate":{"patch":{"summary":"Update Post Manual","description":"Manual override by moderator. Updates the status and triggers metric calculation.","operationId":"update_post_manual_posts__post_id__moderate_patch","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}},{"name":"correct_label","in":"query","required":true,"schema":{"type":"string","title":"Correct Label"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/metrics":{"get":{"summary":"Get Latest Metrics","operationId":"get_latest_metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/Metric"},{"type":"null"}],"title":"Response Get Latest Metrics Metrics Get"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Metric":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"accuracy":{"type":"number","title":"Accuracy","default":0.0},"precision":{"type":"number","title":"Precision","default":0.0},"recall":{"type":"number","title":"Recall","default":0.0},"total_samples":{"type":"integer","title":"Total Samples","default":0},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","title":"Metric"},"Post":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"content":{"type":"string","title":"Content"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"user_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"User Id"},"status":{"type":"string","title":"Status","default":"PENDING"},"toxicity_score":{"type":"number","title":"Toxicity Score","default":0.0},"misinformation_score":{"type":"number","title":"Misinformation Score","default":0.0},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"moderated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Moderated At"},"manual_override":{"type":"boolean","title":"Manual Override","default":false},"correct_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Label"}},"type":"object","required":["content"],"title":"Post"},"User":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"username":{"type":"string","title":"Username"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["username"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}